+++ R/guzioR-package.R
@@ -3,7 +3,9 @@
 
 #' Internal package methods
 #'
+#' @import ggplot2
 #' @import palettes
+#' @import dplyr
 #' @keywords internal
 #' @name guzioR