+++ data-raw/guzior-palettes.R
@@ -21,20 +21,12 @@ plot(b10.pal)
 
 usethis::use_data(b10.pal, overwrite = TRUE)
 
-# Colorblind-Friendly----
-cb.pal <- pal_palette(
-  threecol = c("#003352","#009dca","#c24658"),
-  cpcols = c("#8B2323", "#EE7600", "#EEC900", "chartreuse3", "#0000FF", "#AB82FF", "#CD6889", "#FFA07A", "#FFFF00", "#228B22", "#AFEEEE", "#DDA0DD", "#EE2C2C", "#CDBE70", "#B0B099","#FFD900", "#32CD32", "maroon4", "cornflowerblue", "darkslateblue","#FFFFE0", "#FFEC8B", "peru", "#668B8B", "honeydew","#A020F0", "grey", "#8B4513", "#191970", "#00FF7F","lemonchiffon","#66CDAA", "#5F9EA0", "#A2CD5A", "#556B2F")
+# Palettes from me or friends----
+guzior.pal <- pal_palette(
+  cpcols = c("#8B2323", "#EE7600", "#EEC900", "chartreuse3", "#0000FF", "#AB82FF", "#CD6889", "#FFA07A", "#FFFF00", "#228B22", "#AFEEEE", "#DDA0DD", "#EE2C2C", "#CDBE70", "#B0B099","#FFD900", "#32CD32", "maroon4", "cornflowerblue", "darkslateblue","#FFFFE0", "#FFEC8B", "peru", "#668B8B", "honeydew","#A020F0", "grey", "#8B4513", "#191970", "#00FF7F","lemonchiffon","#66CDAA", "#5F9EA0", "#A2CD5A", "#556B2F"),
+  paired6.col = c("#33a02c", "#c8f9b5", "#e31a1c", "#fccfcf", "#1f78b4", "#a4c9e7", "#e6550d", "#ffcf9c", "#756bb1", "#c3c3e0", "#252525", "#f7f7f7"),
+  threecol = c("#003352","#009dca","#c24658")
   )
-
-plot(cb.pal)
-
-usethis::use_data(cb.pal, overwrite = TRUE)
-
-# All palettes----
-guzior.pal <- c(cb.pal,
-                b10.pal
-                )
 
 plot(guzior.pal)